Rearrange some models.
+ Create one directory for Slot related items.
