--- NLayerArchitecture/ClinicPlatformDTOs/ClinicModels/ClinicSlotInfoModel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ClinicPlatformDTOs.ClinicModels
-{
-    public class ClinicSlotInfoModel
-    {
-        public int Id { get; set; }
-        public TimeOnly StartTime { get; set; }
-        public TimeOnly EndTime { get; set; }
-    }
-}
