+++ NLayerArchitecture/ClinicPlatformDTOs/SlotModels/ClinicSlotRegistrationModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ClinicPlatformDTOs.ClinicModels.Registration
+namespace ClinicPlatformDTOs.SlotModels
 {
     public class ClinicSlotRegistrationModel
     {
