+++ NLayerArchitecture/ClinicPlatformDTOs/SlotModels/SlotInfoModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ClinicPlatformDTOs.SlotModels
+{
+    public class SlotInfoModel
+    {
+        public int? Id { get; set; }
+        public TimeOnly? StartTime { get; set; }
+        public TimeOnly? EndTime { get; set; }
+    }
+}
